Refactor code into package and add requirements
