+++ afvalkalender/__init__.py
@@ -1,0 +1,6 @@
+"""Utilities for fetching waste collection data and exporting calendars."""
+
+from .fetcher import WasteFetcher
+from .ical import export_ical
+
+__all__ = ["WasteFetcher", "export_ical"]
